update issue 33
Status: In Progress

Fixed a bug resulted due to changing the type of the displacement, velocity and acceleration lists from int to float

+++  terasoft-12/Mechanect/Mechanect/Mechanect/Classes/Game.cs
@@ -78,8 +78,8 @@ namespace Mechanect.Classes
                 Graph = new PerformanceGraph();
                 List<string> Commands = new List<string>();
                 List<double> CommandsTime = new List<double>();
-                List<int> Player1Displacement = new List<int>();
-                List<int> Player2Displacement = new List<int>();
+                List<float> Player1Displacement = new List<float>();
+                List<float> Player2Displacement = new List<float>();
                 //initiating testing values 
                 Commands.Add("constantAcceleration");
                 CommandsTime.Add(4);
